requirements.txt
Django==3.2.23
django_cors_headers==4.1.0
djangorestframework_jwt==1.11.0
pymysql==1.1.0
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.24.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
django-gmailapi-backend==0.3.2
drf-jwt==1.19.2
djangorestframework-simplejwt==5.3.1
pyotp==2.9.0